Problem: environment awareness
In certain cases, being aware if there are other tasks, and whether any
of them is queued for execution helps making decisions -- like waiting
for a message from a queue instead of polling.

Solution: introduce `tasks()` and `queued_tasks()` API

While implementing this, I ran into an issue: one of the tests was
locking when running in a true single-threaded environment. The issue
was that another test was leaving an unresolved task, so
`run(None)` was never finishing. This prompted addition of `evict_all()`
to facilitate test execution. I thought that it might have some
practical value in other cases so I made it public.