Problem: single-threaded token counter can be exhausted
Since tokens are `usize` numbers and the counter is a simple increment,
once it reaches `usize::MAX`, can overflow and panic.

Solution: provide clarity that we always need a wrapping increment

This still means tokens can be reused, but hopefully this number of
tasks is not easy to achieve. If this is not true, a better token
mechanism will need to be devised.