Problem: running futures to completion
Within sync environments, there is a need to run all tasks until a
completion of a certain future. Problem is, these futures often get
non-static lifetime.

Solution: implement `single_threaded::block_on` API

This call takes non-static lifetime futures and *lies* (transmutes)
about its lifetime. The hope is that this is actually ok because we know
that the future will be complete by the end of `block_on`, leaving no
trace of it when it's done.