Problem: run_cooperatively marked as `unsafe`
Now, I am not 100% sure it *is* safe -- simply because I've only spent a
few days on this library.

I've published it as `unsafe` at first because I do remember transmuting
some futures into a static lifetime somewhere higher up in the code :)

But the reality is that `spawn` itself requires task's future to be
of static lifetime. Which is how it is in `wasm-bindgen-future`. Which
is how it *should be*.

The only time we're actually extending lifetime is when we block on a
future to run it to completion under the assumption that since it will
run to completion at that point, it is safe to do so (whether I am
actually correct is something to be seen)

Solution: remove the `unsafe` marker

It doesn't 100% guarantee I am right, but this is the best I can come up
with right now.