Problem: yield_animation_frame having an enclosed future
It's really confusing why should it have one. It was done to re-use the
infrastucture that supported `yield_timeout`, which supports this to
effectively allow running async APIs to completion before returning into
the executor.

Solution: separate it from `yield_timeout` and remove the parameter

This future will now return the high-resolution timer, "indicating the
point in time when requestAnimationFrame() starts to execute callback
functions."