Problem: executor won't stop if given control of thread
When executor exited because `until` condition has been satisfied, if
there has been a cooperative future (like `YieldTimeout` or
`YieldAnimationFrame`), it will be started again by one of those
futures, should they have their promises resolved.

This is not the expected behaviour.

Solution: maintain state of satisfaction the exit condition
and don't run the executor if restarted when it has been established
that the exit condition has been met.