Problem: FAQ implies that using async APIs is cumbersome
However, this is not really true. As long as you run a background task
that repeatedly yields to the browser, something like this:

```
executor::spawn(async {
        loop {
            executor::yield_async(futures::future::ready(())).await;
        }
});
```

then async APIs can be used naturally, without any wrapping.

Solution: amend the FAQ to make it a bit clearer that this is not a problem