Problem: block_on API returns an Option
This is confusing (why would it be an Option?).

Solution: make it return the value

It looks like we can pretty much guarantee that the receiver stays
non-dropped during the entire lifetime of `block_on`, hence sending a
oneshot message from the spawned task will be successful and the
receiver will be able to retrieve the data.